FSM Refactor
