+++ dnsconfd/fsm/__init__.py
@@ -1,6 +1,7 @@
 """ Subpackage containing classes managing execution flow """
 
 from .context_event import ContextEvent
-from .context_state import ContextState
 from .exit_code import ExitCode
+from .shared_container import SharedContainer
+from .context_state import ContextState
 from .dnsconfd_context import DnsconfdContext
